+++ src/app/shared/component/loading-modal/loading-modal.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-loading-modal',
+  templateUrl: './loading-modal.component.html'
+})
+export class LoadingModalComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}